modificati i metodi per la sincronizzazione

+++ src/main/java/it/prova/gestionetriage/controller/api/DottoreRestController.java
@@ -16,11 +16,16 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.reactive.function.client.WebClient;
 
+import it.prova.gestionetriage.dto.DottoreDTO;
+import it.prova.gestionetriage.dto.DottoreRequestDTO;
+import it.prova.gestionetriage.dto.DottoreResponseDTO;
 import it.prova.gestionetriage.model.Dottore;
 import it.prova.gestionetriage.service.DottoreService;
-
+import reactor.core.publisher.Mono;
 
 @RestController
 @RequestMapping(value = "/dottore", produces = { MediaType.APPLICATION_JSON_VALUE })
@@ -28,6 +33,9 @@ public class DottoreRestController {
 
 	@Autowired
 	private DottoreService dottoreService;
+
+	@Autowired
+	private WebClient webClient;
 
 	@GetMapping("/{idInput}")
 	public Dottore getDottore(@PathVariable(required = true) Long idInput) {
@@ -49,14 +57,22 @@ public class DottoreRestController {
 		return new ResponseEntity<Page<Dottore>>(results, new HttpHeaders(), HttpStatus.OK);
 	}
 
-	@PostMapping
-	public Dottore createNewDottore(@RequestBody Dottore dottoreInput) {
-		return dottoreService.save(dottoreInput);
-	}
-
 	@PutMapping("/{id}")
 	public Dottore updateDottore(@RequestBody Dottore dottoreInput, @PathVariable Long id) {
+
+		if (dottoreInput.getId() != null)
+			throw new RuntimeException("Non è ammesso fornire un id per la creazione");
+
 		Dottore dottoreToUpdate = dottoreService.get(id);
+
+		ResponseEntity<DottoreResponseDTO> response = webClient.post().uri("")
+				.body(Mono.just(new DottoreRequestDTO(dottoreInput.getId(), dottoreInput.getNome(),
+						dottoreInput.getCognome(), dottoreInput.getCodiceDipendente())), DottoreRequestDTO.class)
+				.retrieve().toEntity(DottoreResponseDTO.class).block();
+
+		if (response.getStatusCode() != HttpStatus.OK)
+			throw new RuntimeException("Errore nella creazione della nuova voce tramite api esterna!!!");
+
 		dottoreToUpdate.setNome(dottoreInput.getNome());
 		dottoreToUpdate.setCognome(dottoreInput.getCognome());
 		dottoreToUpdate.setCodiceDipendente(dottoreInput.getCodiceDipendente());
@@ -65,7 +81,35 @@ public class DottoreRestController {
 
 	@DeleteMapping("/{id}")
 	public void deleteDottore(@PathVariable(required = true) Long id) {
+
+		ResponseEntity<DottoreResponseDTO> response = webClient.post().uri("")
+				.body(Mono.just(new DottoreRequestDTO(id)), DottoreRequestDTO.class).retrieve()
+				.toEntity(DottoreResponseDTO.class).block();
+
+		if (response.getStatusCode() != HttpStatus.OK)
+			throw new RuntimeException("Errore nella creazione della nuova voce tramite api esterna!!!");
+
 		dottoreService.delete(dottoreService.get(id));
+	}
+
+//	#############################################################################
+
+	@PostMapping
+	@ResponseStatus(HttpStatus.CREATED)
+	public DottoreDTO createNew(@RequestBody DottoreDTO dottoreInput) {
+		if (dottoreInput.getId() != null)
+			throw new RuntimeException("Non è ammesso fornire un id per la creazione");
+
+		ResponseEntity<DottoreResponseDTO> response = webClient.post().uri("")
+				.body(Mono.just(new DottoreRequestDTO(dottoreInput.getNome(), dottoreInput.getCognome(),
+						dottoreInput.getCodiceDipendente())), DottoreRequestDTO.class)
+				.retrieve().toEntity(DottoreResponseDTO.class).block();
+
+		if (response.getStatusCode() != HttpStatus.CREATED)
+			throw new RuntimeException("Errore nella creazione della nuova voce tramite api esterna!!!");
+
+		Dottore dottoreInserito = dottoreService.inserisciNuovo(dottoreInput.buildDottoreModel());
+		return DottoreDTO.buildDottoreDTOFromModel(dottoreInserito);
 	}
 
 }
